+++ src/appsettings.h
@@ -11,11 +11,11 @@ public:
     explicit AppSettings(QObject *parent = nullptr);
 
     Q_INVOKABLE int    getInt(const QString &key, int    defaultValue = 0)    const;
-    Q_INVOKABLE float  getFloat(const QString &key, float  defaultValue = 0.f) const;
+    Q_INVOKABLE double getFloat(const QString &key, double defaultValue = 0.0) const;
     Q_INVOKABLE QString getString(const QString &key, const QString &defaultValue = {}) const;
 
     Q_INVOKABLE void setInt(const QString &key, int    value);
-    Q_INVOKABLE void setFloat(const QString &key, float  value);
+    Q_INVOKABLE void setFloat(const QString &key, double value);
     Q_INVOKABLE void setString(const QString &key, const QString &value);
     Q_INVOKABLE void save();
